remove zip if it already existed

+++ ZipBuilder/Sources/ZipBuilder/FileManager+Utils.swift
@@ -84,12 +84,12 @@ public extension FileManager {
     return cacheRoot
   }
 
-  /// Removes a directory if it exists. This is helpful to clean up error handling for checks that
+  /// Removes a directory or file if it exists. This is helpful to clean up error handling for checks that
   /// shouldn't fail. The only situation this could potentially fail is permission errors or if a
   /// folder is open in Finder, and in either state the user needs to close the window or fix the
   /// permissions. A fatal error will be thrown in those situations.
-  func removeDirectoryIfExists(at url: URL) {
-    guard directoryExists(at: url) else { return }
+  func removeIfExists(at url: URL) {
+    guard directoryExists(at: url) || fileExists(atPath: url.path) else { return }
 
     do {
       try removeItem(at: url)
